data_ingestion_handler/requirements.txt
boto3
langchain
openai
langchain-core
langchain-community
langchain-openai
langchain-text-splitters
faiss-cpu